pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Task manager"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "asgiref>=3.9.1",
    "autopep8>=2.3.2",
    "certifi>=2025.8.3",
    "charset-normalizer>=3.4.3",
    "dj-database-url>=3.0.1",
    "django==4.2.24",
    "django-bootstrap5>=25.2",
    "django-filter>=25.1",
    "dotenv>=0.9.9",
    "exceptiongroup>=1.3.0",
    "flake8>=7.3.0",
    "gunicorn>=23.0.0",
    "idna>=3.10",
    "iniconfig>=2.1.0",
    "isort>=6.0.1",
    "mccabe>=0.7.0",
    "packaging>=25.0",
    "pluggy>=1.6.0",
    "psycopg2-binary>=2.9.10",
    "pycodestyle>=2.14.0",
    "pyflakes>=3.4.0",
    "pygments>=2.19.2",
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "rollbar>=1.3.0",
    "sqlparse>=0.5.3",
    "tomli>=2.2.1",
    "typing-extensions>=4.15.0",
    "urllib3>=2.5.0",
]

[build-system] 
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel] 
packages = ["task_manager"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "task_manager.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
